BC-1951 Add getFileUrl to video and audio browser command
