+++ static/scripts/ckeditor/plugins/ckeditor5-filebrowser/src/image-browser-command.js
@@ -1,8 +1,6 @@
 import Command from '@ckeditor/ckeditor5-core/src/command';
-import getCookie from '../../../../helpers/cookieManager';
+import FileBrowserHelper from './file-browser-helper';
 import createFilebrowserModal from './file-browser-modal';
-
-const apiV3BasePath = '/api/v3';
 
 export default class ImageBrowserCommand extends Command {
 	execute() {
@@ -11,17 +9,7 @@ export default class ImageBrowserCommand extends Command {
 		const dialogTitle = this.editor.t('Image Properties');
 
 		const onCreate = async () => {
-			const topicId = $('#content-blocks').data('topicid');
-			const schoolId = $('#content-blocks').data('schoolid');
-			const courseFileUrl = document.getElementById('url-input').value;
-
-			let imageUrl;
-
-			if (topicId !== undefined && schoolId !== undefined) {
-				imageUrl = await this.copyAsLessonFile(schoolId, 'lessons', topicId, courseFileUrl);
-			} else {
-				imageUrl = courseFileUrl;
-			}
+			const imageUrl = await FileBrowserHelper.getFileUrl();
 
 			if (!imageUrl) return;
 			const imageAltText = document.getElementById('alt-text-input').value;
@@ -37,22 +25,5 @@ export default class ImageBrowserCommand extends Command {
 			});
 		};
 		createFilebrowserModal(this.editor, this.editor.t, dialogTitle, onCreate, additionalInput);
-	}
-
-	async copyAsLessonFile(schoolId, parentType, parentId, url) {
-		const regex = new RegExp('(?<=name=).+', 'g');
-		const fileNameMatch = url.match(regex);
-
-		if (!fileNameMatch) return;
-
-		const fileRecord = await $.ajax(`${apiV3BasePath}/file/upload-from-url/${schoolId}/${parentType}/${parentId}`, {
-			method: 'POST',
-			headers: {
-				Authorization: `Bearer ${getCookie('jwt')}`,
-			},
-			data: { url: `${window.location.origin}${url}`, fileName: `${fileNameMatch[0]}` },
-		});
-
-		return `/api/v3/file/download/${fileRecord.id}/${fileRecord.name}`;
 	}
 }
